refactored to extract console IO and file processing logic to separate classes
